Spark
Work exclusively with Spark. Only the SparkContext is needed
The lowest and highest temperatures measured each year for the period 1950-2014. Provide the lists sorted in descending order with respect to the maximum temperature.

The average monthly temperature for each available station in Sweden.

A list of stations with their associated maximum measured temperatures and maximum measured daily precipitation

The average monthly precipitation for the Östergotland region (list of stations is provided in the separate file) for the period 1993-2016.